facture_maroc_with_ai/facture_maroc-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
pdfkit==1.0.0
jinja2==3.1.2
